refactor web view and create TabbyCardView
