+++ Sources/Tabby/SwiftUI/Constants/BaseURL.swift
@@ -25,8 +25,17 @@ public enum BaseURL {
         }
         
         public enum Widgets {
-            public static var url: String {
+            public static var promo: String {
                 return "https://widgets.tabby.ai/tabby-promo.html"
+            }
+
+            public static var card: String {
+                return "https://widgets.tabby.ai/tabby-card.html"
+            }
+
+            @available(*, deprecated, renamed: "promo")
+            public static var url: String {
+                return promo
             }
         }
