+++ Sources/Tabby/SwiftUI/Extensions/Dictionary+QueryString.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+extension Dictionary where Key == String {
+
+    var queryString: String {
+        map { "\($0.key)=\($0.value)" }
+            .joined(separator: "&")
+    }
+}
